+++ Mods ChoGGi/Minimap/metadata.lua
@@ -1,0 +1,18 @@
+return PlaceObj("ModDef", {
+	"title", "Minimap v0.1",
+	"version", 1,
+	"saved", 1542888000,
+	"image", "Preview.png",
+	"id", "ChoGGi_Minimap",
+	"steam_id", "1571476937",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Minimap.lua",
+		"Code/Script.lua",
+	},
+	"description", [[It's an image of the topography you can click on to move the camera around (useful for setting transport routes).
+
+This is an image, so click target isn't perfect.
+Next update will (hopefully) be clickable icons of domes, rovers, etc.]],
+})